+++ composition-js/src/merging/merge.ts
@@ -34,6 +34,7 @@ import {
   Subgraphs,
   JOIN_VERSIONS,
   TAG_VERSIONS,
+  INACCESSIBLE_VERSIONS,
   NamedSchemaElement,
   executableDirectiveLocations,
   errorCauses,
@@ -50,12 +51,16 @@ import {
   joinStrings,
   FederationMetadata,
   printSubgraphNames,
-  Subgraph,
   federationIdentity,
   linkIdentity,
   coreIdentity,
   FEDERATION_OPERATION_TYPES,
   LINK_VERSIONS,
+  federationMetadata,
+  FeatureDefinition,
+  Subgraph,
+  errorCode,
+  withModifiedErrorNodes,
 } from "@apollo/federation-internals";
 import { ASTNode, GraphQLError, DirectiveLocation } from "graphql";
 import {
@@ -80,7 +85,6 @@ import {
 
 const linkSpec = LINK_VERSIONS.latest();
 const joinSpec = JOIN_VERSIONS.latest();
-const tagSpec = TAG_VERSIONS.latest();
 
 export type MergeResult = MergeSuccess | MergeFailure;
 
@@ -161,25 +165,16 @@ function isGraphQLBuiltInDirective(def: DirectiveDefinition): boolean {
   return !!def.schema().builtInDirective(def.name);
 }
 
-function isMergedDirective(definition: DirectiveDefinition | Directive): boolean {
-  // If it's a directive application, then we skip it unless it's a graphQL built-in
-  // (even if the definition itself allows executable locations, this particular
-  // application is an type-system element and we don't want to merge it).
-  if (definition instanceof Directive) {
-    // We have special code in `Merger.prepareSupergraph` to include the _definition_ of @tag in the supergraph
-    // because we want to make sure it's included as a "core feature", and so we want this method
-    // to return `false` for `@tag` definition, but we *do* want to merge `@tag` applications so we
-    // special case it.
-    // Note that this is a temporary solution: a more principled way to have directive propagated
-    // is coming and will remove the hard-coding.
-    return definition.name === 'tag' || isGraphQLBuiltInDirective(definition.definition!);
-  } else if (isGraphQLBuiltInDirective(definition)) {
-    // We never "merge" graphQL built-in definitions, since they are built-in and
-    // don't need to be defined.
-    return false;
-  }
-  return definition.locations.some(loc => executableDirectiveLocations.includes(loc));
+type MergedDirectiveInfo = {
+  specInSupergraph: FeatureDefinition,
+  definitionInSubgraph: (subgraph: Subgraph) => DirectiveDefinition,
 }
+
+const MERGED_FEDERATION_DIRECTIVES: MergedDirectiveInfo[] = [
+  { specInSupergraph: TAG_VERSIONS.latest(), definitionInSubgraph: (subgraph) => subgraph.metadata().tagDirective() },
+  { specInSupergraph: INACCESSIBLE_VERSIONS.latest(), definitionInSubgraph: (subgraph) => subgraph.metadata().inaccessibleDirective() },
+];
+
 // Access the type set as a particular root in the provided `SchemaDefinition`, but ignoring "query" type
 // that only exists due to federation operations. In other words, if a subgraph don't have a query type,
 // but one was automatically added for _entities and _services, this method returns 'undefined'.
@@ -221,10 +216,6 @@ function typeKindToString(t: NamedType): string {
   return t.kind.replace("Type", " Type");
 }
 
-function hasTagUsage(subgraph: Subgraph): boolean {
-  return subgraph.metadata().tagDirective().applications().length > 0;
-}
-
 function locationString(locations: DirectiveLocation[]): string {
   if (locations.length === 0) {
     return "";
@@ -239,6 +230,7 @@ class Merger {
   readonly hints: CompositionHint[] = [];
   readonly merged: Schema = new Schema();
   readonly subgraphNamesToJoinSpecName: Map<string, string>;
+  readonly mergedFederationDirectiveNames = new Set<string>();
 
   constructor(readonly subgraphs: Subgraphs, readonly options: CompositionOptions) {
     this.names = subgraphs.names();
@@ -251,13 +243,45 @@ class Merger {
     // pass a `as` to the methods below if necessary. However, as we currently don't propagate any subgraph directives to
     // the supergraph outside of a few well-known ones, we don't bother yet.
     linkSpec.addToSchema(this.merged);
-    linkSpec.applyFeatureToSchema(this.merged, joinSpec, undefined, 'EXECUTION');
+    const errors = linkSpec.applyFeatureToSchema(this.merged, joinSpec, undefined, 'EXECUTION');
+    assert(errors.length === 0, "We shouldn't have errors adding the join spec to the (still empty) supergraph schema");
 
-    if (this.subgraphs.values().some(hasTagUsage)) {
-      linkSpec.applyFeatureToSchema(this.merged, tagSpec);
+    for (const mergedInfo of MERGED_FEDERATION_DIRECTIVES) {
+      this.validateAndMaybeAddSpec(mergedInfo);
     }
 
     return joinSpec.populateGraphEnum(this.merged, this.subgraphs);
+  }
+
+  private validateAndMaybeAddSpec({specInSupergraph, definitionInSubgraph}: MergedDirectiveInfo) {
+    let nameInSupergraph: string | undefined;
+    for (const subgraph of this.subgraphs) {
+      const directive = definitionInSubgraph(subgraph);
+      if (directive.applications().length === 0) {
+        continue;
+      }
+
+      if (!nameInSupergraph) {
+        nameInSupergraph = directive.name;
+      } else if (nameInSupergraph !== directive.name) {
+        this.reportMismatchError(
+          ERRORS.LINK_IMPORT_NAME_MISMATCH,
+          `The federation "@${specInSupergraph.url.name}" directive is imported with mismatched name between subgraphs: it is imported as `,
+          subgraph.metadata().federationFeature()?.directive!,
+          this.subgraphs.values().map((s) => s.metadata().federationFeature()?.directive),
+          (linkDef) => `"@${federationMetadata(linkDef.schema())?.federationFeature()?.directiveNameInSchema(specInSupergraph.url.name)}"`,
+        );
+        return;
+      }
+    }
+
+    // If we get here with `nameInSupergraph` unset, it means there is no usage for the directive at all and we
+    // don't bother adding the spec to the supergraph.
+    if (nameInSupergraph) {
+      this.mergedFederationDirectiveNames.add(nameInSupergraph);
+      const errors = linkSpec.applyFeatureToSchema(this.merged, specInSupergraph, nameInSupergraph === specInSupergraph.url.name ? undefined : nameInSupergraph);
+      assert(errors.length === 0, "We shouldn't have errors adding the join spec to the (still empty) supergraph schema");
+    }
   }
 
   private joinSpecName(subgraphIndex: number): string {
@@ -266,6 +290,25 @@ class Merger {
 
   private metadata(idx: number): FederationMetadata {
     return this.subgraphs.values()[idx].metadata();
+  }
+
+  private isMergedDirective(definition: DirectiveDefinition | Directive): boolean {
+    // If it's a directive application, then we skip it unless it's a graphQL built-in
+    // (even if the definition itself allows executable locations, this particular
+    // application is an type-system element and we don't want to merge it).
+    if (definition instanceof Directive) {
+      // We have special code in `Merger.prepareSupergraph` to include the _definition_ of merged federation
+      // directives in the supergraph, so we don't have to merge those _definition_, but we *do* need to merge
+      // the applications.
+      // Note that this is a temporary solution: a more principled way to have directive propagated
+      // is coming and will remove the hard-coding.
+      return this.mergedFederationDirectiveNames.has(definition.name) || isGraphQLBuiltInDirective(definition.definition!);
+    } else if (isGraphQLBuiltInDirective(definition)) {
+      // We never "merge" graphQL built-in definitions, since they are built-in and
+      // don't need to be defined.
+      return false;
+    }
+    return definition.locations.some(loc => executableDirectiveLocations.includes(loc));
   }
 
   merge(): MergeResult {
@@ -329,10 +372,13 @@ class Merger {
           // to subgraphs appropriately). and then simply _assert_ that `Schema.validate()` doesn't throw as a sanity
           // check.
           this.merged.validate();
+          // Lastly, we validate that the API schema of the supergraph can be successfully compute, which currently will surface issues around
+          // misuses of `@inaccessible` (there should be other errors in theory, but if there is, better find it now rather than later).
+          this.merged.toAPISchema();
         } catch (e) {
           const causes = errorCauses(e);
           if (causes) {
-            this.errors.push(...causes);
+            this.errors.push(...this.updateInaccessibleErrorsWithLinkToSubgraphs(causes));
           } else {
             // Not a GraphQLError, so probably a programing error. Let's re-throw so it can be more easily tracked down.
             throw e;
@@ -380,7 +426,7 @@ class Merger {
     // bit round-about, it's a tad simpler code-wise to do this way.
     for (const subgraph of this.subgraphsSchema) {
       for (const directive of subgraph.allDirectives()) {
-        if (!isMergedDirective(directive)) {
+        if (!this.isMergedDirective(directive)) {
           continue;
         }
         if (!this.merged.directive(directive.name)) {
@@ -401,7 +447,7 @@ class Merger {
     );
   }
 
-  private reportMismatchError<TMismatched extends SchemaElement<any, any>>(
+  private reportMismatchError<TMismatched extends { sourceAST?: ASTNode }>(
     code: ErrorCodeDefinition,
     message: string,
     mismatchedElement:TMismatched,
@@ -424,7 +470,7 @@ class Merger {
     );
   }
 
-  private reportMismatchErrorWithSpecifics<TMismatched extends SchemaElement<any, any>>(
+  private reportMismatchErrorWithSpecifics<TMismatched extends { sourceAST?: ASTNode }>(
     code: ErrorCodeDefinition,
     message: string,
     mismatchedElement: TMismatched,
@@ -452,7 +498,7 @@ class Merger {
     );
   }
 
-  private reportMismatchHint<TMismatched extends SchemaElement<any, any>>(
+  private reportMismatchHint<TMismatched extends { sourceAST?: ASTNode }>(
     hintId: HintID,
     message: string,
     supergraphElement: TMismatched,
@@ -483,7 +529,7 @@ class Merger {
     );
   }
 
-  private reportMismatch<TMismatched extends SchemaElement<any, any>>(
+  private reportMismatch<TMismatched extends { sourceAST?: ASTNode }>(
     supergraphElement:TMismatched,
     subgraphElements: (TMismatched | undefined)[],
     mismatchAccessor: (element: TMismatched, isSupergraph: boolean) => string | undefined,
@@ -823,7 +869,7 @@ class Merger {
         // Note that if we change our mind on this semantic and wanted directives on external to propagate, then we'll also
         // need to update the merging of fields since external fields are filtered out (by this very method).
         for (const directive of source.appliedDirectives) {
-          if (isMergedDirective(directive)) {
+          if (this.isMergedDirective(directive)) {
             // Contrarily to most of the errors during merging that "merge" errors for related elements, we're logging one
             // error for every application here. But this is because there error is somewhat subgraph specific and is
             // unlikely to span multiple subgraphs. In fact, we could almost have thrown this error during subgraph validation
@@ -1380,7 +1426,7 @@ class Merger {
     //   definition is the intersection of all definitions (meaning that if there divergence in
     //   locations, we only expose locations that are common everywhere).
     this.mergeDescription(sources, dest);
-    if (sources.some((s) => s && isMergedDirective(s))) {
+    if (sources.some((s) => s && this.isMergedDirective(s))) {
       this.mergeExecutionDirectiveDefinition(sources, dest);
     }
   }
@@ -1573,7 +1619,7 @@ class Merger {
     for (const source of sources) {
       if (source) {
         for (const directive of source.appliedDirectives) {
-          if (isMergedDirective(directive)) {
+          if (this.isMergedDirective(directive)) {
             names.add(directive.name);
           }
         }
@@ -1728,5 +1774,31 @@ class Merger {
         }
       }
     }
+  }
+
+  private updateInaccessibleErrorsWithLinkToSubgraphs(errors: GraphQLError[]): GraphQLError[] {
+    return errors.map((e) => {
+      if (errorCode(e) !== ERRORS.REFERENCED_INACCESSIBLE.code) {
+        return e;
+      }
+
+      const reference = e.extensions['inaccessible_reference'];
+      const referenceNodes = reference && typeof reference === 'string'
+        ? sourceASTs(...this.subgraphsSchema.map((schema) => schema.elementByCoordinate(reference)))
+        : [];
+      const element = e.extensions['inaccessible_element'];
+      // We only point to subgraphs where the element is marked inaccesssible, as other subgraph are not relevant.
+      const elementNodes = element && typeof element === 'string'
+        ? sourceASTs(
+            ...this.subgraphsSchema
+              .map((schema) => schema.elementByCoordinate(element))
+              .filter((elt) => elt?.hasAppliedDirective(federationMetadata(elt.schema())!.inaccessibleDirective()))
+          )
+        : [];
+
+      // Note that the error we get will likely have some AST nodes, but those will be to the supergraph and that's not
+      // useful to the user so we completely discard them.
+      return withModifiedErrorNodes(e, [...referenceNodes, ...elementNodes]);
+    });
   }
 }