+++ internals-js/src/inaccessibleSpec.ts
@@ -1,6 +1,7 @@
 import { FeatureDefinition, FeatureDefinitions, FeatureUrl, FeatureVersion } from "./coreSpec";
 import {
   DirectiveDefinition,
+  ErrGraphQLValidationFailed,
   FieldDefinition,
   isCompositeType,
   isInterfaceType,
@@ -9,21 +10,30 @@ import {
 } from "./definitions";
 import { GraphQLError, DirectiveLocation } from "graphql";
 import { registerKnownFeature } from "./knownCoreFeatures";
+import { ERRORS } from "./error";
+import { createDirectiveSpecification } from "./directiveAndTypeSpecification";
 
 export const inaccessibleIdentity = 'https://specs.apollo.dev/inaccessible';
+
+export const inaccessibleLocations = [
+  DirectiveLocation.FIELD_DEFINITION,
+  DirectiveLocation.OBJECT,
+  DirectiveLocation.INTERFACE,
+  DirectiveLocation.UNION,
+];
+
+export const inaccessibleDirectiveSpec = createDirectiveSpecification({
+  name: 'inaccessible',
+  locations: [...inaccessibleLocations],
+});
 
 export class InaccessibleSpecDefinition extends FeatureDefinition {
   constructor(version: FeatureVersion) {
     super(new FeatureUrl(inaccessibleIdentity, 'inaccessible', version));
   }
 
-  addElementsToSchema(schema: Schema) {
-    this.addDirective(schema, 'inaccessible').addLocations(
-      DirectiveLocation.FIELD_DEFINITION,
-      DirectiveLocation.OBJECT,
-      DirectiveLocation.INTERFACE,
-      DirectiveLocation.UNION,
-    );
+  addElementsToSchema(schema: Schema): GraphQLError[] {
+    return this.addDirectiveSpec(schema, inaccessibleDirectiveSpec);
   }
 
   inaccessibleDirective(schema: Schema): DirectiveDefinition<Record<string, never>> {
@@ -63,6 +73,7 @@ export function removeInaccessibleElements(schema: Schema) {
     );
   }
 
+  const errors = [];
   for (const type of schema.types()) {
     // @inaccessible can only be on composite types.
     if (!isCompositeType(type)) {
@@ -76,10 +87,16 @@ export function removeInaccessibleElements(schema: Schema) {
         // field type will be `undefined`).
         if (reference.kind === 'FieldDefinition') {
           if (!reference.hasAppliedDirective(inaccessibleDirective)) {
-            throw new GraphQLError(
-              `Field ${reference.coordinate} returns an @inaccessible type without being marked @inaccessible itself.`,
-              reference.sourceAST,
-            );
+            // We ship the inaccessible type and it's invalid reference in the extensions so composition can extract those and add proper links
+            // in the subgraphs for those elements.
+            errors.push(ERRORS.REFERENCED_INACCESSIBLE.err({
+              message: `Field "${reference.coordinate}" returns @inaccessible type "${type}" without being marked @inaccessible itself.`,
+              nodes: reference.sourceAST,
+              extensions: {
+                "inaccessible_element": type.coordinate,
+                "inaccessible_reference": reference.coordinate,
+              }
+            }));
           }
         }
         // Other references can be:
@@ -91,5 +108,9 @@ export function removeInaccessibleElements(schema: Schema) {
       const toRemove = (type.fields() as FieldDefinition<any>[]).filter(f => f.hasAppliedDirective(inaccessibleDirective));
       toRemove.forEach(f => f.remove());
     }
+  }
+
+  if (errors.length > 0) {
+    throw ErrGraphQLValidationFailed(errors, `Schema has ${errors.length === 1 ? 'an invalid use' : 'invalid uses'} of the @inaccessible directive.`);
   }
 }