+++ query-planner-js/CHANGELOG.md
@@ -5,8 +5,11 @@ This CHANGELOG pertains only to Apollo Federation packages in the 2.x range. The
 ## vNEXT
 
 > The changes noted within this `vNEXT` section have not been released yet.  New PRs and commits which introduce changes should include an entry in this `vNEXT` section as part of their development.  When a release is being prepared, a new header will be (manually) created below and the appropriate changes within that release will be moved into the new section.
+- Adds Support for `@tag/v0.2`, which allows the `@tag` directive to be additionally placed on arguments, scalars, enums, enum values, input objects, and input object fields. [PR #1652](https://github.com/apollographql/federation/pull/1652).
 
-- _Nothing yet! Stay tuned!_
+## v2.0.0-preview.8
+
+- Released in sync with other federation packages but no changes to this package.
 
 ## v2.0.0-preview.7